pyproject.toml
[project]
name = "job-recommendation-system"
version = "0.1.0"
description = "A job recommendation and skill assessment system based on graph learning and LLM"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "torch>=2.0.0",
    "torch-geometric>=2.4.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "langgraph>=0.0.30",
    "neo4j>=5.20.0",
    "faiss-cpu>=1.7.0",
    "pydantic>=2.5.0",
    "typing-extensions>=4.8.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
index = "https://pypi.org/simple"

[tool.uv.package]
exclude = ["tests", "test_*.py", "*_test.py"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88